+++ xknx/cemi/cemi_handler.py
@@ -11,7 +11,14 @@ import asyncio
 import logging
 from typing import TYPE_CHECKING
 
-from xknx.exceptions import CommunicationError, ConfirmationError, ConversionError
+from xknx.exceptions import (
+    CommunicationError,
+    ConfirmationError,
+    ConversionError,
+    DataSecureError,
+)
+from xknx.secure.data_secure import DataSecure
+from xknx.secure.keyring import Keyring
 from xknx.telegram import IndividualAddress, Telegram, TelegramDirection, tpci
 
 from .cemi_frame import CEMIFrame
@@ -21,6 +28,7 @@ if TYPE_CHECKING:
     from xknx.xknx import XKNX
 
 logger = logging.getLogger("xknx.cemi")
+data_secure_logger = logging.getLogger("xknx.data_secure")
 
 # See 3/6/3 EMI_IMI §4.1.5 Data Link Layer messages
 REQUEST_TO_CONFIRMATION_TIMEOUT = 3
@@ -32,7 +40,15 @@ class CEMIHandler:
     def __init__(self, xknx: XKNX) -> None:
         """Initialize CEMIHandler class."""
         self.xknx = xknx
+        self.data_secure: DataSecure | None = None
         self._l_data_confirmation_event = asyncio.Event()
+
+    def data_secure_init(self, keyring: Keyring | None) -> None:
+        """Initialize DataSecure."""
+        if keyring is None:
+            self.data_secure = None
+        else:
+            self.data_secure = DataSecure.init_from_keyring(keyring)
 
     async def send_telegram(self, telegram: Telegram) -> None:
         """Create a CEMIFrame from a Telegram and send it to the CEMI Server."""
@@ -43,8 +59,11 @@ class CEMIHandler:
                 self.xknx.current_address if telegram.source_address.raw == 0 else None
             ),
         )
-        self._l_data_confirmation_event.clear()
         logger.debug("Outgoing CEMI: %s", cemi)
+        if self.data_secure is not None:
+            cemi = self.data_secure.outgoing_cemi(cemi=cemi)
+
+        self._l_data_confirmation_event.clear()
         try:
             await self.xknx.knxip_interface.send_cemi(cemi)
         except (ConversionError, CommunicationError) as ex:
@@ -74,6 +93,13 @@ class CEMIHandler:
             logger.warning("Received unexpected L_DATA_REQ frame: %s", cemi)
             return
 
+        # TODO: do we have to decrypt Data Secure L_DATA_CON?
+        if self.data_secure is not None:
+            try:
+                cemi = self.data_secure.received_cemi(cemi=cemi)
+            except DataSecureError as err:
+                data_secure_logger.warning("Could not decrypt CEMI frame: %s", err)
+                return
         # TODO: remove telegram init from CEMIFrame class and move it here?
         telegram = cemi.telegram
         telegram.direction = TelegramDirection.INCOMING