+++ xknx/telegram/apci.py
@@ -15,6 +15,7 @@ from typing import ClassVar, cast
 
 from xknx.dpt import DPTArray, DPTBinary
 from xknx.exceptions import ConversionError
+from xknx.secure.data_secure_asdu import SecureData, SecurityControlField
 from xknx.telegram.address import IndividualAddress
 
 
@@ -94,6 +95,9 @@ class APCIExtendedService(Enum):
     INDIVIDUAL_ADDRESS_SERIAL_READ = 0x03DC
     INDIVIDUAL_ADDRESS_SERIAL_RESPONSE = 0x03DD
     INDIVIDUAL_ADDRESS_SERIAL_WRITE = 0x03DE
+
+    # DataSecure
+    APCI_SEC = 0x03F1
 
 
 class APCI(ABC):
@@ -205,6 +209,8 @@ class APCI(ABC):
                 return IndividualAddressSerialResponse.from_knx(raw)
             if apci == APCIExtendedService.INDIVIDUAL_ADDRESS_SERIAL_WRITE.value:
                 return IndividualAddressSerialWrite.from_knx(raw)
+            if apci == APCIExtendedService.APCI_SEC.value:
+                return SecureAPDU.from_knx(raw)
 
         raise ConversionError(f"Class not implemented for APCI {apci:#012b}.")
 
@@ -1605,3 +1611,35 @@ class IndividualAddressSerialWrite(APCI):
     def __str__(self) -> str:
         """Return object as readable string."""
         return f'<IndividualAddressSerialWrite serial="{self.serial.hex()}" address="{self.address}" />'
+
+
+class SecureAPDU(APCI):
+    """SecureAPDU service."""
+
+    CODE = APCIExtendedService.APCI_SEC
+
+    def __init__(self, scf: SecurityControlField, secured_data: SecureData) -> None:
+        """Initialize a new instance of AuthorizeRequest."""
+        self.scf = scf
+        self.secured_data = secured_data
+
+    def calculated_length(self) -> int:
+        """Get length of APCI payload."""
+        return 2 + len(self.secured_data)
+
+    @classmethod
+    def from_knx(cls, raw: bytes) -> SecureAPDU:
+        """Parse/deserialize from KNX/IP raw data."""
+        return cls(
+            scf=SecurityControlField.from_knx(raw[2]),
+            secured_data=SecureData.from_knx(raw[3:]),
+        )
+
+    def to_knx(self) -> bytearray:
+        """Serialize to KNX/IP raw data."""
+        payload = self.scf.to_knx() + self.secured_data.to_knx()
+        return encode_cmd_and_payload(self.CODE, appended_payload=payload)
+
+    def __str__(self) -> str:
+        """Return object as readable string."""
+        return f'<SecureAPDU scf="{self.scf}" secured_data={self.secured_data!r} />'